Implement Homework1 form
I've used Flask-WTforms to add a form to solve the UMUC Homework 1
problem.  The changes are that in __init__.py, I added a secret key.  In
views, I created a WTForm class and code in the '/hw1' endpoint to solve
the problem.  In the hw1.html template, I added code to render the form
and output.